lab04/src/workers/requirements.txt
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
asyncpg==0.29.0
greenlet==3.0.1
requests==2.31.0
furl==2.1.3
pgvector==0.2.1
aiokafka==0.8.1
faust-streaming==0.10.0
kafka-python==2.0.2
sentence_transformers==2.2.2
python-json-logger==2.0.4
